Applying non-blocking write to pipe_stdout/pipe_stderr
The original code only made the reading end of the pipe non-blocking,
as it needs to empty the pipe after the sandbox.
But, if the student's program actually writes a lot on the writing end
of the pipe, the pipe will block and cause the program to timeout.
It is better to actually report this error (too much output)
rather than waiting for the task to end.
In both cases, the output will be truncated and it will not pass the tests,
as we may expect the problem author to not require the student to output
more than 65536 bytes or so of data on stdout or on stderr.